- незначительно скорректировал AdController - добавил поля для связи сущностей в AdEntity, UserEntity - скорректировал тип идентификатора в AdRepository, UserRepository - добавил CRUD-методы в AdService, AdServiceImpl (кроме updateComment) - добавил запросы к БД в CommentRepository - добавил маппер ResponseWrapperCommentMapper и класс ResponseWrapperCommentMapperImpl - добавил структуру таблиц проекта в WiKi проекта

+++ src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -3,6 +3,5 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.model.UserEntity;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 }
